+++ packages/genesys/packages/react-components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Versions
 
+## 3.2.4
+
+[Package.json] `react-components-core` update (not_equal_validator)
+
+## 3.2.3
+
+[Package.json] `react-components-core` update (useNumericInput)
+
 ## 3.2.2
 
 [useColor] Moved to react-components-core